Change return type of countByExample from int to long

+++ core/mybatis-generator-systests-ibatis2-java2/src/test/java/mbg/test/ib2j2/hierarchical/HierarchicalJava2Test.java
@@ -211,7 +211,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             FieldsonlyExample example = new FieldsonlyExample();
             example.createCriteria().andIntegerfieldGreaterThan(new Integer(5));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -383,7 +383,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             PkonlyExample example = new PkonlyExample();
             example.createCriteria().andIdGreaterThan(new Integer(4));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example = new PkonlyExample();
@@ -1096,7 +1096,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             PkfieldsExample example = new PkfieldsExample();
             example.createCriteria().andLastnameLike("J%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1386,7 +1386,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             PkblobsExample example = new PkblobsExample();
             example.createCriteria().andIdLessThan(new Integer(4));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1771,7 +1771,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             PkfieldsblobsExample example = new PkfieldsblobsExample();
             example.createCriteria().andId1NotEqualTo(new Integer(3));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1958,7 +1958,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             FieldsblobsExample example = new FieldsblobsExample();
             example.createCriteria().andFirstnameLike("S%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -2976,7 +2976,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             AwfulTableExample example = new AwfulTableExample();
             example.createCriteria().andEMailLike("fred@%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
