+++ core/mybatis-generator-systests-ibatis2-java5/src/test/java/mbg/test/ib2j5/conditional/ConditionalJava5Test.java
@@ -225,7 +225,7 @@ public class ConditionalJava5Test extends AbstractConditionalJava5Test {
 
             FieldsonlyExample example = new FieldsonlyExample();
             example.createCriteria().andIntegerfieldGreaterThan(5);
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -403,7 +403,7 @@ public class ConditionalJava5Test extends AbstractConditionalJava5Test {
 
             PkonlyExample example = new PkonlyExample();
             example.createCriteria().andIdGreaterThan(4);
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -1123,7 +1123,7 @@ public class ConditionalJava5Test extends AbstractConditionalJava5Test {
 
             PkfieldsExample example = new PkfieldsExample();
             example.createCriteria().andLastnameLike("J%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1414,7 +1414,7 @@ public class ConditionalJava5Test extends AbstractConditionalJava5Test {
 
             PkblobsExample example = new PkblobsExample();
             example.createCriteria().andIdLessThan(4);
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1812,7 +1812,7 @@ public class ConditionalJava5Test extends AbstractConditionalJava5Test {
 
             PkfieldsblobsExample example = new PkfieldsblobsExample();
             example.createCriteria().andId1NotEqualTo(3);
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -2005,7 +2005,7 @@ public class ConditionalJava5Test extends AbstractConditionalJava5Test {
 
             FieldsblobsExample example = new FieldsblobsExample();
             example.createCriteria().andFirstnameLike("S%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -3027,7 +3027,7 @@ public class ConditionalJava5Test extends AbstractConditionalJava5Test {
 
             AwfulTableExample example = new AwfulTableExample();
             example.createCriteria().andEMailLike("fred@%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
